+++ Embraced/WordListTrialsViewController.swift
@@ -16,6 +16,7 @@ import AVKit
 
 class WordListTrialsViewController: ActiveStepViewController {
 
+    let TEST_NAME = "WordList"
     var index = 1
     var isPlaying = false
     var isRecording = false
@@ -270,56 +271,38 @@ class WordListTrialsViewController: ActiveStepViewController {
     
     private func externalStorage() {
         SVProgressHUD.show()
-        let filePath = "\(FirebaseStorageManager.shared.pid!)/WordList/\(soundFileName)"
+        let filePath = "\(FirebaseStorageManager.shared.pid!)/\(TEST_NAME)/\(soundFileName)"
         
         if Utility.fileExist(filePath) {
-            let storage = Storage.storage()
-            let storageRef = storage.reference()
-            let participantRef = storageRef.child(filePath)
-            
-            let uploadTask = participantRef.putFile(from: recordedAudioURL, metadata: nil) { (metadata, error) in
+            FirebaseStorageManager.shared.externalStorage(filePath: filePath, fileUrl: recordedAudioURL) { (uploadTask, error) in
                 if error != nil {
-                    print("Error: \(error?.localizedDescription)")
-                    SVProgressHUD.showError(withStatus: "An error has happened")
+                    SVProgressHUD.showError(withStatus: error?.localizedDescription)
                 }
                 
-                switch (self.index) {
-                    case 1:
-                        WordListModel.shared.task_1 = filePath
-                        break
-                    case 2:
-                        WordListModel.shared.task_2 = filePath
-                        break
-                    case 3:
-                        WordListModel.shared.task_3 = filePath
-                        break
-                    case 4:
-                        WordListModel.shared.task_4 = filePath
-                        break
-                    case 5:
-                        WordListModel.shared.task_5 = filePath
-                        break
-                    case 6:
-                        WordListModel.shared.interference = filePath
-                        break
-                    case 7:
-                        WordListModel.shared.shortTerm = filePath
-                        break
-                    default: break
+                if let task = uploadTask {
+                    task.observe(.success, handler: { (snapshot) in
+                        switch (self.index) {
+                        case 1: WordListModel.shared.task_1 = filePath
+                        case 2: WordListModel.shared.task_2 = filePath
+                        case 3: WordListModel.shared.task_3 = filePath
+                        case 4: WordListModel.shared.task_4 = filePath
+                        case 5: WordListModel.shared.task_5 = filePath
+                        case 6: WordListModel.shared.interference = filePath
+                        case 7: WordListModel.shared.shortTerm = filePath
+                        default: break
+                        }
+                        
+                        FirebaseStorageManager.shared.addDataToDocument(payload: [
+                            "wordList": WordListModel.shared.printModel()
+                        ])
+                        
+                        self.nextBtn.isHidden = false
+                        
+                        // Delete file from device
+                        Utility.deleteFile(filePath)
+                        SVProgressHUD.dismiss()
+                    })
                 }
-                
-                FirebaseStorageManager.shared.addDataToDocument(payload: [
-                    "wordList": WordListModel.shared.printModel()
-                ])
-
-            }
-            
-            uploadTask.observe(.success) { (snapshot) in
-                self.nextBtn.isHidden = false
-                
-                // Delete file from device
-                Utility.deleteFile(filePath)
-                SVProgressHUD.dismiss()
             }
         }
     }
